requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
tensorflow==2.15.0
keras==2.15.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
matplotlib==3.8.3
scikit-learn==1.4.1.post1
gunicorn==21.2.0
python-dotenv==1.0.1
flask-sqlalchemy==3.1.1
flask-login==0.6.3
authlib==1.3.0
requests==2.31.0
